Eliminate unnecessary boxing
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@1059210 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/area/Block.java
@@ -130,7 +130,7 @@ public class Block extends BlockParent {
      */
     public int getStartIndent() {
         Integer startIndent = (Integer)getTrait(Trait.START_INDENT);
-        return (startIndent != null ? startIndent.intValue() : 0);
+        return (startIndent != null ? startIndent : 0);
     }
 
 }
